refactor: separated callbacks from app.py
